requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
dnspython>=2.4.0
builtwith>=1.3.4
wappalyzer>=0.3.1
colorama>=0.4.6
python-nmap>=0.7.1
validators>=0.22.0
lxml>=4.9.3
pyOpenSSL>=23.3.0

# New dependencies for SEO, Content & Products features
extruct>=0.16.0
feedparser>=6.0.10
tldextract>=5.1.0
PyYAML>=6.0.1
